test(ngldm): move the non-IBSI matrix check out of the IBSI oracle file
test_2d_ngldm_ibsi.h held two NGLD-matrix checks. The first compares against IBSI's published
Fig. 3.19 matrix and belongs there. The second does not: its ground truth is the NGLDM worked out in
a StackOverflow answer -- test_data.h names the URL beside the image -- and it runs with IBSI mode
OFF, which is the mode the IBSI definition does not describe. It was still called
test_2d_ngldm_matrix_correctness_nonibsi_mode_ibsi(), so the file it sat in and the _ibsi suffix both
claimed an oracle the assertion never uses.

It moves to test_2d_ngldm_regression.h as
test_2d_ngldm_matrix_correctness_nonibsi_mode_regression(). Regression rather than an oracle test
because the reference cannot be regenerated from this tree: a forum post is not a tool that can be
run, there is no version to record and no generator to write (SPEC 6.4). It still earns its place --
it holds the non-IBSI matrix, which keys on unique grey tones rather than the 0..max IBSI range, to
the shape a second implementation arrived at, so a change in the dependence counting is caught.
Making it an oracle claim means running MATLAB or Octave, pinning what that produces, and shipping
the generator beside it; the report says so where the move is recorded.

<vector> comes off test_2d_ngldm_ibsi.h in passing: test_ref_vals.h, which the file already includes,
supplies it -- the same rule already applied to the other five family branches.

No values change and no registry row moves; the matrix checks back no feature row.

Verified: runAllTests 802/802 on Windows and 798/798 under ASan+UBSan on Linux (the four missing
there are the DICOM Hounsfield loader mechanics, compiled out by NOEXTRAS), no sanitizer report;
check_coverage.py --check, check_test_names.py --check and audit/scan_ngldm_coverage.py --check all
clean, and the audit artifact regenerates unchanged.
