+++ tests/test_all.cc
@@ -2088,11 +2088,6 @@ TEST(TEST_NYXUS, TEST_2D_NGLDM_MATRIX_CORRECTNESS_IBSI)
 	ASSERT_NO_THROW (test_2d_ngldm_matrix_correctness_ibsi());
 }
 
-TEST(TEST_NYXUS, TEST_2D_NGLDM_MATRIX_CORRECTNESS_NONIBSI_MODE_IBSI)
-{
-	ASSERT_NO_THROW(test_2d_ngldm_matrix_correctness_nonibsi_mode_ibsi());
-}
-
 TEST(TEST_NYXUS, TEST_2D_NGLDM_LDE_IBSI)
 {
 	ASSERT_NO_THROW(test_2d_ngldm_lde_ibsi());
@@ -2156,6 +2151,11 @@ TEST(TEST_NYXUS, TEST_2D_NGLDM_DCNUN_IBSI)
 TEST(TEST_NYXUS, TEST_2D_NGLDM_DCP_IBSI)
 {
 	ASSERT_NO_THROW(test_2d_ngldm_dcp_ibsi());
+}
+
+TEST(TEST_NYXUS, TEST_2D_NGLDM_MATRIX_CORRECTNESS_NONIBSI_MODE_REGRESSION)
+{
+	ASSERT_NO_THROW(test_2d_ngldm_matrix_correctness_nonibsi_mode_regression());
 }
 
 TEST(TEST_NYXUS, TEST_2D_NGLDM_GLM_REGRESSION)
